refactor hook names to use query
